# resources/files/blog_backend__swagger.json
{"openapi":"3.0.2","info":{"title":"Rest API Specification","description":"API Base path is `/api/` and it is authorized via temporary Bearer token after login.","version":"1.0.0"},"paths":{"/api/articles/{id}":{"get":{"tags":["Articles"],"summary":"Get detailed view of existing single article.","parameters":[{"name":"id","in":"path","description":"User account ID","required":true,"schema":{"type":"integer"}}],"responses":{"200":{"description":"Success - json","content":{"application/json":{"schema":{"type":"object","properties":{"id":{"type":"integer"},"title":{"type":"string"},"content":{"type":"string"},"author":{"type":"object","properties":{"id":{"type":"integer"},"email":{"type":"string"},"name":{"type":"string"},"role":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"articles":{"type":"array"}}},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}}}}}},"400":{"description":"Bad request"},"401":{"description":"Unauthorized"},"404":{"description":"Not found"},"406":{"description":"Not acceptable"},"500":{"description":"Internal server error"}}},"put":{"tags":["Articles"],"summary":"Update existing article.","parameters":[{"name":"id","in":"path","description":"Article ID","required":true,"schema":{"type":"integer"}}],"requestBody":{"description":"ArticleEntityDto update DTO.","content":{"application/json":{"schema":{"type":"object","properties":{"title":{"nullable":true,"type":"string"},"content":{"nullable":true,"type":"string"}}}}},"required":true},"responses":{"200":{"description":"Success - json"},"400":{"description":"Bad request"},"401":{"description":"Unauthorized"},"404":{"description":"Not found"},"406":{"description":"Not acceptable"},"500":{"description":"Internal server error"}}},"delete":{"tags":["Articles"],"summary":"Permanently delete existing article.","parameters":[{"name":"id","in":"path","description":"User account ID","required":true,"schema":{"type":"integer"}}],"responses":{"200":{"description":"Success - json"},"400":{"description":"Bad request"},"401":{"description":"Unauthorized"},"404":{"description":"Not found"},"406":{"description":"Not acceptable"},"500":{"description":"Internal server error"}}}},"/api/articles":{"get":{"tags":["Articles"],"summary":"Get list of all articles in database.","responses":{"200":{"description":"Success - json"},"400":{"description":"Bad request"},"401":{"description":"Unauthorized"},"404":{"description":"Not found"},"406":{"description":"Not acceptable"},"500":{"description":"Internal server error"}}},"post":{"tags":["Articles"],"summary":"Create new article..","requestBody":{"description":"ArticleEntityDto creation DTO.","content":{"application/json":{"schema":{"type":"object","properties":{"title":{"nullable":true,"type":"string"},"content":{"nullable":true,"type":"string"}}}}},"required":true},"responses":{"200":{"description":"Success - json","content":{"application/json":{"schema":{"type":"object","properties":{"id":{"type":"integer"},"title":{"type":"string"},"content":{"type":"string"},"author":{"type":"object","properties":{"id":{"type":"integer"},"email":{"type":"string"},"name":{"type":"string"},"role":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"articles":{"type":"array"}}},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}}}}}},"400":{"description":"Bad request"},"401":{"description":"Unauthorized"},"404":{"description":"Not found"},"406":{"description":"Not acceptable"},"500":{"description":"Internal server error"}}}},"/api/auth/login":{"post":{"tags":["Auth"],"summary":"Login via user account credentials to acquire temporary token.","responses":{"200":{"description":"Success - json"},"500":{"description":"Internal server error"}}}},"/api/auth/register":{"post":{"tags":["Auth"],"summary":"User account register for blog.","responses":{"200":{"description":"Success - json"},"500":{"description":"Internal server error"}}}},"/api/users/{id}":{"get":{"tags":["Users"],"summary":"Get detailed view of existing single user account.","parameters":[{"name":"id","in":"path","description":"User account ID","required":true,"schema":{"type":"integer"}}],"responses":{"200":{"description":"Success - json","content":{"application/json":{"schema":{"type":"object","properties":{"id":{"type":"integer"},"email":{"type":"string"},"name":{"type":"string"},"role":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"articles":{"type":"array"}}}}}},"400":{"description":"Bad request"},"401":{"description":"Unauthorized"},"404":{"description":"Not found"},"406":{"description":"Not acceptable"},"500":{"description":"Internal server error"}}},"put":{"tags":["Users"],"summary":"Update existing user account.","parameters":[{"name":"id","in":"path","description":"User account ID","required":true,"schema":{"type":"integer"}}],"requestBody":{"description":"UserEntityDto update DTO.","content":{"application/json":{"schema":{"type":"object","properties":{"email":{"nullable":true,"type":"string"},"name":{"nullable":true,"type":"string"},"role":{"nullable":true,"type":"string"}}}}},"required":true},"responses":{"200":{"description":"Success - json"},"400":{"description":"Bad request"},"401":{"description":"Unauthorized"},"404":{"description":"Not found"},"406":{"description":"Not acceptable"},"500":{"description":"Internal server error"}}},"delete":{"tags":["Users"],"summary":"Permanently delete existing user.","parameters":[{"name":"id","in":"path","description":"User account ID","required":true,"schema":{"type":"integer"}}],"responses":{"200":{"description":"Success - json"},"400":{"description":"Bad request"},"401":{"description":"Unauthorized"},"404":{"description":"Not found"},"406":{"description":"Not acceptable"},"500":{"description":"Internal server error"}}}},"/api/users":{"get":{"tags":["Users"],"summary":"Get list of all users accounts in database.","responses":{"200":{"description":"Success - json"},"400":{"description":"Bad request"},"401":{"description":"Unauthorized"},"404":{"description":"Not found"},"406":{"description":"Not acceptable"},"500":{"description":"Internal server error"}}},"post":{"tags":["Users"],"summary":"Create new user account.","requestBody":{"description":"UserEntityDto creation DTO.","content":{"application/json":{"schema":{"type":"object","properties":{"email":{"nullable":true,"type":"string"},"password":{"nullable":true,"type":"string"},"name":{"nullable":true,"type":"string"},"role":{"nullable":true,"type":"string"}}}}},"required":true},"responses":{"200":{"description":"Success - json","content":{"application/json":{"schema":{"type":"object","properties":{"id":{"type":"integer"},"email":{"type":"string"},"name":{"type":"string"},"role":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"articles":{"type":"array"}}}}}},"400":{"description":"Bad request"},"401":{"description":"Unauthorized"},"404":{"description":"Not found"},"406":{"description":"Not acceptable"},"500":{"description":"Internal server error"}}}}}}